Weekly DevToMasterMerge

+++ angular/src/app/components/chatroom-list/chatroom-list.component.ts
@@ -1,8 +1,6 @@
-import { UserStoreService } from './../../services/user-store.service';
 import { ChatService } from './../../services/chat.service';
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from 'src/app/services/auth.service';
-import { Router } from '@angular/router';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 
 
@@ -16,8 +14,6 @@ export class ChatroomListComponent implements OnInit {
   currentUser = this.authService.getCurrentUser();
   userChatsAsUserA$: any[] = [];
   userChatsAsUserB$: any[] = [];
-  userChatsAsUserALength: number;
-  userChatsAsUserBLength: number;
 
   faTimes = faTimes;
 
@@ -30,37 +26,26 @@ export class ChatroomListComponent implements OnInit {
   constructor(
     private authService: AuthService,
     private chatService: ChatService,
-    private router: Router,
-    private userStoreService: UserStoreService
   ) {
+    // initialize all the chatrooms from the current user, where the current user ID is equal to userA
     this.chatService.getAllChatroomsAsUserA().subscribe(chatrooms => {
-      this.userChatsAsUserALength = chatrooms.length;
       this.userChatsAsUserA$ = chatrooms;
-    }, error => {
-      console.log('Error in profile-page - TODO delete this console.log() before finishing WebProg!');
+    }, () => {
+      console.log('Error in chatroom-list-page - TODO delete this console.log() before finishing WebProg!');
     });
+
+    // initialize all the chatrooms from the current user, where the current user ID is equal to userB
     this.chatService.getAllChatroomsAsUserB().subscribe(chatrooms => {
-      this.userChatsAsUserBLength = chatrooms.length;
       this.userChatsAsUserB$ = chatrooms;
-    }, error => {
-      console.log('Error in profile-page - TODO delete this console.log() before finishing WebProg!');
+    }, () => {
+      console.log('Error in chatroom-list-page - TODO delete this console.log() before finishing WebProg!');
     });
   }
 
   ngOnInit() {
   }
 
-  // deleteChatroom() {
-  //   this.chatService.deleteChatroom(this.roomToBeDeleted).subscribe(() => {
-  //     if (this.userChatsAsUserALength === 1 || this.userChatsAsUserBLength === 1) {
-  //       this.router.navigateByUrl('/', { skipLocationChange: true }).then(() =>
-  //         this.router.navigate(['/chats']));
-  //     }
-  //   });
-  //   this.closeModal();
-  // }
-
-  // TODO doesn't work properly
+  // delete chatroom from database
   deleteChatroom() {
     let indexNumber: number;
     if (this.chatsArrayName === 'userChatsAsUserA$') {
@@ -72,7 +57,7 @@ export class ChatroomListComponent implements OnInit {
         }
         this.userChatsAsUserA$.splice(indexNumber, 1);
       });
-    } else {
+    } else if (this.chatsArrayName === 'userChatsAsUserB$') {
       this.chatService.deleteChatroom(this.roomToBeDeleted).subscribe(() => {
         for (let index = 0; index < this.userChatsAsUserB$.length; index++) {
           if (this.userChatsAsUserB$[index].id === this.roomToBeDeleted) {
@@ -86,7 +71,6 @@ export class ChatroomListComponent implements OnInit {
   }
 
   openModal(chatroomId: string, chatsArrayName: string) {
-    console.log('id: ' + chatroomId);
     this.modalIsOpen = true;
     this.display = 'block';
     this.roomToBeDeleted = chatroomId;
